% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrence.R
\name{concurrence}
\alias{concurrence}
\title{Concurrence Analysis}
\usage{
concurrence(
  data,
  bin = c(40, 60, 80, 120, 160, 200, 250, 300, 350, 400),
  ref,
  cgm,
  group_var,
  trans = FALSE
)
}
\arguments{
\item{data}{Paired Data Set}

\item{bin}{Default is 40,60,80,120,160,200,250,300,350,400. Users can provide their own}

\item{ref}{Provide Reference Readings}

\item{cgm}{Provide Sensor Readings such as Gl or ANA}

\item{group_var}{Any Categorical Variables}

\item{trans}{Transpose table. Default is FALSE}
}
\value{
Gt Table
}
\description{
Concurrence Analysis
}
