% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_error_grid.R
\name{consensus_error_grid}
\alias{consensus_error_grid}
\title{Consensus Error Grid Analysis}
\usage{
consensus_error_grid(
  df,
  ref,
  cgm,
  unit = "gram",
  group_var = NULL,
  show_plot = FALSE,
  title = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{df}{Paired Data Set}

\item{ref}{Provide Reference Readings}

\item{cgm}{Provide Sensor Readings such as Gl or ANA}

\item{unit}{Either "gram" or "mol"}

\item{group_var}{Any Categorical Variables. Default is NULL}

\item{show_plot}{Either TRUE or FALSE. Default is FALSE.}

\item{title}{Title of Plot}

\item{xlab}{xlab of Plot}

\item{ylab}{ylab of Plot}
}
\value{
GT Table or GGplot2
}
\description{
Consensus Error Grid Analysis
}
