% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sensor_serial_number_to_UID.R
\name{convert_sensor_serial_number_to_UID}
\alias{convert_sensor_serial_number_to_UID}
\title{Convert Sensor Serial Number to UID}
\usage{
convert_sensor_serial_number_to_UID(x)
}
\arguments{
\item{x}{Column of Sensor Serial Number you would like to convert to UID}
}
\value{
A Vector
}
\description{
Convert Sensor Serial Number to UID
}
