% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_UID_to_sensor_serial_number.R
\name{convert_UID_to_sensor_serial_number}
\alias{convert_UID_to_sensor_serial_number}
\title{Convert UID to Sensor Serial Number}
\usage{
convert_UID_to_sensor_serial_number(x)
}
\arguments{
\item{x}{Column of UID you would like to convert to Sensor Serial Number}
}
\value{
A Vector
}
\description{
Convert UID to Sensor Serial Number
}
