% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices_readings.R
\name{devices_readings}
\alias{devices_readings}
\title{Import devices.csv and readings.csv upload data}
\usage{
devices_readings(devices_path, readings_path, index = NULL)
}
\arguments{
\item{devices_path}{File path of devices.csv.}

\item{readings_path}{File path of readings.csv.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import devices.csv and readings.csv upload data
}
