% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexcom.R
\name{dexcom}
\alias{dexcom}
\title{Import Dexcom csv or xlsx upload data}
\usage{
dexcom(dexcom_path, index = NULL)
}
\arguments{
\item{dexcom_path}{File path of dexcom}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import Dexcom csv or xlsx upload data
}
