% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_measure.R
\name{diff_measure}
\alias{diff_measure}
\title{Difference Measures Analysis}
\usage{
diff_measure(data, reference, group_var)
}
\arguments{
\item{data}{Paired Data Set}

\item{reference}{Either TRUE or FALSE}

\item{group_var}{Any Categorical Variables}
}
\value{
Gt Table
}
\description{
Difference Measures Analysis
}
