% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{events}
\alias{events}
\title{Import Events csv upload data}
\usage{
events(events_path, index = NULL)
}
\arguments{
\item{events_path}{file path of events.csv}

\item{index}{index to get an individual or specific upload}
}
\value{
A tibble
}
\description{
Import Events csv upload data
}
