% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_case_when.R
\name{fct_case_when}
\alias{fct_case_when}
\title{Factor case_when}
\usage{
fct_case_when(...)
}
\arguments{
\item{...}{A sequence of two-sided formulas consistent with dplyr::case_when}
}
\value{
A vector factor
}
\description{
Factor case_when
}
