% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_path.R
\name{filter_path}
\alias{filter_path}
\title{Filter matched patterns from path}
\usage{
filter_path(file_path, filter_text = "^$|pattern")
}
\arguments{
\item{file_path}{File path of events.csv, gluc.csv, glucPlus.csv, and freestyle.csv.}

\item{filter_text}{Match a fixed string that you would like to remove. Default is an empty string.}
}
\value{
A list
}
\description{
Filter matched patterns from path
}
