% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freestyle.R
\name{freestyle}
\alias{freestyle}
\title{Import Freestyle csv upload data}
\usage{
freestyle(freestyle_path, index = NULL)
}
\arguments{
\item{freestyle_path}{File path of freestyle.csv.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import Freestyle csv upload data
}
