% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ysi.R
\name{import_ysi}
\alias{import_ysi}
\title{Import YSI csv upload data}
\usage{
import_ysi(ysi_path, index = NULL)
}
\arguments{
\item{ysi_path}{File path of YSI.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import YSI csv upload data
}
