% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ketone.R
\name{ketone}
\alias{ketone}
\title{Import Ketone csv upload data}
\usage{
ketone(ketone_path, index = NULL)
}
\arguments{
\item{ketone_path}{File path of ketone.csv.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import Ketone csv upload data
}
