% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobi.R
\name{mobi}
\alias{mobi}
\title{Import Mobi csv upload data}
\usage{
mobi(events, gluc, index = NULL)
}
\arguments{
\item{events}{File path of events.csv.}

\item{gluc}{File path of gluc.csv.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import Mobi csv upload data
}
