% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobi_anaPlus.R
\name{mobi_anaPlus}
\alias{mobi_anaPlus}
\title{Import anaPlus.csv or ana.csv upload data}
\usage{
mobi_anaPlus(events, ana, index = NULL)
}
\arguments{
\item{events}{File path of events.csv.}

\item{ana}{File path of anaPlus.csv or ana.csv.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import anaPlus.csv or ana.csv upload data
}
