% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneMinutes.R
\name{oneminutes}
\alias{oneminutes}
\title{Import oneMinutes csv upload data}
\usage{
oneminutes(events, one_minutes, index = NULL)
}
\arguments{
\item{events}{File path of events.csv.}

\item{one_minutes}{File path of oneMinutes.csv.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import oneMinutes csv upload data
}
