% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairBg.R
\name{PairBg}
\alias{PairBg}
\title{Pair BG Reference}
\usage{
PairBg(sensor_data, apol = NULL)
}
\arguments{
\item{sensor_data}{Sensor Data}

\item{apol}{Boolean. Is sensor Apol or Mobi}
}
\value{
Tibble
}
\description{
Pair BG Reference
}
