% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edc.R
\name{read_edc}
\alias{read_edc}
\title{Read EDC Data}
\usage{
read_edc(directory, name_of_edc)
}
\arguments{
\item{directory}{Directory of EDC}

\item{name_of_edc}{Name of EDC data}
}
\value{
A tibble
}
\description{
Read EDC Data
}
