% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_normal.R
\name{round_normal}
\alias{round_normal}
\title{Fix Round Half to Even}
\usage{
round_normal(x, digits)
}
\arguments{
\item{x}{A Numeric Vector}

\item{digits}{Integer Indicating the Number of Decimal Places}
}
\value{
A Numeric Vector
}
\description{
Fix Round Half to Even
}
\examples{
x <- c(0.125,0.215)
round_normal(x,2)
}
