% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibonics.R
\name{sibonics}
\alias{sibonics}
\title{Import Sibionics Sensor Data}
\usage{
sibonics(sibonics_path, index = NULL)
}
\arguments{
\item{sibonics_path}{File path of xlsx or csv}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import Sibionics Sensor Data
}
