% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_agreement.R
\name{system_agreement}
\alias{system_agreement}
\title{System Agreement Analysis}
\usage{
system_agreement(
  data,
  type = "Glucose",
  breakpoint = 70,
  ref,
  cgm,
  unit = "mg/dL",
  group_var,
  transpose = FALSE,
  wider_name = NULL
)
}
\arguments{
\item{data}{Paired Data Set}

\item{type}{Either Glucose or Ketone. Case is not sensitive}

\item{breakpoint}{It depends on which type and unit. Default is 70}

\item{ref}{Provide Reference Readings}

\item{cgm}{Provide Sensor Readings such as Gl or ANA}

\item{unit}{Either mg/dL or mmol/L for glucose but mmol/L for ketone. Default is mg/dL}

\item{group_var}{Any Categorical Variables}

\item{transpose}{Transpose table. Default is FALSE}

\item{wider_name}{If transpose is True, please provide the transposed column}
}
\value{
Gt Table
}
\description{
System Agreement Analysis
}
