% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atna.R
\name{atna}
\alias{atna}
\title{Import Atna csv upload data}
\usage{
atna(events, gluc, index = NULL)
}
\arguments{
\item{events}{File path of events.csv.}

\item{gluc}{File path of gluc.csv.}

\item{index}{Index to get an individual or specific upload. Default is NULL.}
}
\value{
A tibble
}
\description{
Import Atna csv upload data
}
