% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sensor_serial_number_to_UID.R
\name{convert_sensor_serial_number_to_UID}
\alias{convert_sensor_serial_number_to_UID}
\title{Convert Sensor Serial Number to UID}
\usage{
convert_sensor_serial_number_to_UID(data, col)
}
\arguments{
\item{data}{Data of interest}

\item{col}{Col of Sensor Serial Number you would like to convert to UID}
}
\value{
A tibble
}
\description{
Convert Sensor Serial Number to UID
}
