% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_UID_to_sensor_serial_number.R
\name{convert_UID_to_sensor_serial_number}
\alias{convert_UID_to_sensor_serial_number}
\title{Convert UID to Sensor Serial Number}
\usage{
convert_UID_to_sensor_serial_number(data, col)
}
\arguments{
\item{data}{Data of interest}

\item{col}{Col of UID you would like to convert to Sensor Serial Number}
}
\value{
A tibble
}
\description{
Convert UID to Sensor Serial Number
}
