% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_measure.R
\name{diff_measure}
\alias{diff_measure}
\title{Difference Measures Analysis}
\usage{
diff_measure(
  data,
  ref,
  cgm,
  group_var,
  unit = "mg/dL",
  reference_breakpoint = TRUE
)
}
\arguments{
\item{data}{Paired Data Set}

\item{ref}{Provide Reference Readings such as YSI or BG}

\item{cgm}{Provide Sensor Readings such as Gl or ANA}

\item{group_var}{Any Categorical Variables}

\item{unit}{Either mg/dL or mmol/L. Default is mg/dL}

\item{reference_breakpoint}{Default is TRUE}
}
\value{
Gt Table
}
\description{
Difference Measures Analysis
}
