% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobi.R
\name{mobi}
\alias{mobi}
\title{Import Mobi csv upload data}
\usage{
mobi(events, gluc, index = NULL)
}
\arguments{
\item{events}{file path of events.csv}

\item{gluc}{file path of gluc.csv}

\item{index}{index to get an individual or specific upload}
}
\value{
A tibble
}
\description{
Import Mobi csv upload data
}
